make agent creator injectable
